update note
no reason to trust the word we're about to run not to mess with the prog
or index and to not pollute the stack for the next word.  Line up the
usize and string to be next so we can pop them off, run the word, then
pop them back on.